Process as many splats as possible in a single render pass
Prior to this patch, two render passes were used per splat which can
become quite slow if many splats are active simoultaneously. This patch
modifies the splat shader such that a predefined number of splats can
be processed in one go, speeding up the splat rendering significantly.

The maximum size of a splat group can be set through the
`splatGroupSize` URL search parameter. The default value is 64, which
should be supported on all WebGL 2 platforms.